fixed make clean
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@27 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/atgpib/init.c
@@ -3,7 +3,7 @@
 #include <board.h>
 
 
-uint16      ibbase = IBBASE;	/* base addr of GPIB interface registers  */
+unsigned int      ibbase = IBBASE;	/* base addr of GPIB interface registers  */
 uint8       ibirq  = IBIRQ;	/* interrupt request line for GPIB (1-7)  */
 uint8       ibdma  = IBDMA;     /* DMA channel                            */
